requirements.txt
Django==3.0
beautifulsoup4==4.9.1
djangorestframework==3.11.0
